+++ src/io.h
@@ -1,3 +1,36 @@
+template <typename T> int8_t sign(T val) {
+  return (T(0) < val) - (val < T(0));
+}
+
+
+void printRange(int r0 = 0, int r1 = 0) {
+  if (r1 == 0)
+    for (byte i = 0; i < r0; i++) {
+      PT(i);
+      PT('\t');
+    }
+  else
+    for (byte i = r0; i < r1; i++) {
+      PT(i);
+      PT('\t');
+    }
+  PTL();
+}
+template <typename T> void printList(T * arr, byte len = DOF) {
+  String temp = "";
+  for (byte i = 0; i < len; i++) {
+    temp += String(int(arr[i]));
+    temp += '\t';
+    //PT((T)(arr[i]));
+    //PT('\t');
+  }
+  PTL(temp);
+}
+template <typename T> void printTable(T * list) {
+  printRange(0, DOF);
+  printList(list, DOF);
+}
+
 char getUserInputChar() {//take only the first character, allow "no line ending", "newline", "carriage return", and "both NL & CR"
   while (!Serial.available());
   char result = Serial.read();
@@ -27,8 +60,9 @@ void read_serial() {
     newCmdIdx = 2;
     token = Serial.read();
 
-    if (token == T_SKILL_DATA)
+    if (token == T_SKILL_DATA) {
       readSerialUntil(dataBuffer, '~');
+    }
     else if (Serial.available() > 0) {
       String cmdBuffer;//may overflow after many iterations. I use this method only to support "no line ending" in the serial monitor
       if (token == T_INDEXED_SIMULTANEOUS_BIN || token == T_LISTED_BIN || token == T_MOVE_BIN || token == T_BEEP_BIN) {
@@ -100,66 +134,3 @@ bool sensorConnectedQ(int n) {
 
 
 //short tools
-
-
-void EEPROMWriteInt(int p_address, int p_value)
-{
-  byte lowByte = ((p_value >> 0) & 0xFF);
-  byte highByte = ((p_value >> 8) & 0xFF);
-  EEPROM.update(p_address, lowByte);
-  EEPROM.update(p_address + 1, highByte);
-  delay(5);
-}
-//This function will read a 2-byte integer from the EEPROM at the specified address and address + 1
-int EEPROMReadInt(int p_address)
-{
-  byte lowByte = EEPROM.read(p_address);
-  byte highByte = EEPROM.read(p_address + 1);
-  return ((lowByte << 0) & 0xFF) + ((highByte << 8) & 0xFF00);
-}
-inline int8_t eeprom(int address, byte x = 0, byte y = 0, byte yDim = 1) {
-  return EEPROM.read(address + x * yDim + y);
-}
-
-void flushEEPROM(char b) {
-  for (int j = 0; j < 1024; j++) {
-    EEPROM.update(j, b);
-  }
-}
-
-void printEPROM() {
-  for (int i = 0; i < 1024; i++) {
-    PT(eeprom(i));
-    PT('\t');
-    if ((i % 16) == 15)
-      PTL();
-  }
-}
-
-template <typename T> void printEEPROMList(int EEaddress, byte len = DOF) {
-  for (byte i = 0; i < len; i++) {
-    PT((T)(EEPROM.read(EEaddress + i)));
-    PT('\t');
-  }
-  PTL();
-}
-
-void saveCalib(int8_t *var) {
-  for (byte i = 0; i < DOF; i++) {
-    EEPROM.update(CALIB + i, var[i]);
-    //    calibratedZeroPosition[i] = EEPROMReadInt(ZERO_POSITIONS + i * 2) + float(var[i]) * eeprom(ROTATION_DIRECTION, i);
-    EEPROMWriteInt(CALIBRATED_ZERO_POSITIONS + i * 2, EEPROMReadInt(ZERO_POSITIONS + i * 2) + float(var[i]) * eeprom(ROTATION_DIRECTION, i));
-  }
-}
-
-void saveMelody(int &address, byte melody[], int len ) {
-  EEPROM.update(address--, len);
-  for (byte i = 0; i < len; i++)
-    EEPROM.update(address --, melody[i]);
-  PTL(address);
-}
-
-
-inline int loadAngleLimit(byte idx, byte dim) {
-  return EEPROMReadInt(ANGLE_LIMIT + idx * 4 + dim * 2);
-}
